+++ src/bonuses/fortune-handler.mjs
@@ -168,7 +168,9 @@ export const handleFortune = (options) => {
     }
 
     if (options.fortuneCount === options.misfortuneCount) {
-        options.dice = '1d20';
+        if (!options.dice) {
+            options.dice = '1d20';
+        }
         return;
     }